requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==37.0.2
ffmpeg==1.4
h11==0.13.0
idna==3.3
multidict==6.0.2
outcome==1.1.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.0.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.27.1
selenium==4.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.2.0
urllib3==1.26.9
wsproto==1.1.0
youtube-dl==2021.12.17
